Levels/Level 2/13-email-sender/requirements.txt
# Email Sender Dependencies
email-validator>=2.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
ruff>=0.1.0

# Gmail API and Google Sheets
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1

# Web UI
flask>=3.0.0

# Environment variables
python-dotenv>=1.0.0

# Optional: Jinja2 for advanced templates
jinja2>=3.1.0
